Fix version mismatch (#121)
In the last release, the version was not actually updated. It should be in the future since the bumper just added to bump the version.

The version mismatch didn't break anything, because this package is not published anywhere, and is installed in aiidalab-docker-stack directly from the git repo based on the git version tag, which was correct. But if you try to identify the version in the current image, for example, the `mamba list` using the version tag from the file.