Simple URL sanitization (#137)
Currently, if the app URL is not found in app metadata,
the `app.url` variable (that we render on the single app page)
is not None, but instead it is a string
that says "Field 'external_url' not present in app metadata".
When this string is rendered as a href target, it results
in an invalid link leading to a 404 page.

Here's we just add a simple validation to check
that the URL starts with http. If the condition does not match the field won't show.

This is a hotfix solution to
https://github.com/aiidalab/aiidalab/issues/329
until the root cause (https://github.com/aiidalab/aiidalab-registry/issues/112) is fixed.