Add strict validation check for core dependencies (#130)
Along with aiidalab/aiidalab#347 this commit fixes aiidalab/aiidalab#318

First of all, after `aiidalab/aiidalab#347` the app manager can check the validity of versions from the app registry w.r.t the core packages. If the core package requirements are not met, in this commit, we will not show the versions that are not compatible with core packages. 

Besides, we add a popped dialog when the version to install is selected which will show the list of dependencies that will be installed or changed. It also gives a warning for the potential danger that installing the app will break another app, or if there are already dependencies compatible exist, it will give the hint of which dependencies cause the incompatible. 

---------

Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>
Co-authored-by: Daniel Hollas <danekhollas@gmail.com>