Add auto-detected home warning message mechanism (#183)
This PR adds a hard-coded path for displaying a notification widget between the home controls and the app accordion. The mechanism detects if `/home/jovyan/.aiidalab/home_app_warning.md` is present, and if so, displays its contents in a `Markdown` widget.

As mentioned, the present mechanism is hard-coded in this implementation. However, a more flexible system (see #54) is soon to be discussed to expand the concept of user notifications beyond the home app and with a more general scope.

Thanks to @danielhollas for assisting with authoring a test :)
---------

Co-authored-by: Daniel Hollas <daniel.hollas@bristol.ac.uk>