Replace `jinja2` with `markdown` for loading the warning message document (#189)
This PR fixes an instability with using `jinja2` and `IPython.display.Markdown` to load the warning message document. Here we instead use `markdown.Markdown` to convert the markdown document to HTML and render it using an `ipywidget.HTML` widget, which is simpler to work with (no `ipywidgets.Output` widgets required!).